Create separate file for key_blob. Fixed things in nodejs_conductor
