+++ hc_dpki/src/lib.rs
@@ -11,6 +11,7 @@ pub mod key_bundle;
 pub mod keypair;
 pub mod password_encryption;
 pub mod utils;
+pub mod key_blob;
 
 pub const SEED_SIZE: usize = 32;
 pub const SIGNATURE_SIZE: usize = 64;
